Refactor XML generation and validation in workflow and code
